Extract purchase modal into its own independant component
