+++ static/js/src/PurchaseModal/components/FreeTrialRadio/FreeTrialRadio.test.tsx
@@ -2,27 +2,22 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { add, format } from "date-fns";
-import { formatter } from "../../../renderers/form-renderer";
-import * as usePreview from "../../APICalls/usePreview";
+import { formatter } from "../../../advantage/subscribe/renderers/form-renderer";
 import { preview } from "../../utils/test/Mocks";
 
 import FreeTrialRadio from "./FreeTrialRadio";
 
 const total = 123456;
 
-jest.spyOn(usePreview, "default").mockImplementation(() => {
-  return {
-    isLoading: false,
-    isError: false,
-    isSuccess: true,
-    data: { ...preview, total: total },
-    error: null,
-  };
-});
+const previewData = { ...preview, total: total };
 
 it("displays a message explaining the trial if free trial is selected", () => {
   render(
-    <FreeTrialRadio isUsingFreeTrial={true} setIsUsingFreeTrial={() => {}} />
+    <FreeTrialRadio
+      preview={previewData}
+      isUsingFreeTrial={true}
+      setIsUsingFreeTrial={() => {}}
+    />
   );
   screen.getByText("Your free trial ends:");
   screen.getByText(
@@ -37,7 +32,11 @@ it("displays a message explaining the trial if free trial is selected", () => {
 
 it("does not display the message if pay now is selected", () => {
   render(
-    <FreeTrialRadio isUsingFreeTrial={false} setIsUsingFreeTrial={() => {}} />
+    <FreeTrialRadio
+      preview={previewData}
+      isUsingFreeTrial={false}
+      setIsUsingFreeTrial={() => {}}
+    />
   );
   expect(screen.queryByText("Your free trial ends:")).not.toBeInTheDocument();
 });
@@ -46,6 +45,7 @@ it("calls the setIsUsingFreeTrial function when the radio is clicked", () => {
   const setIsUsingFreeTrial = jest.fn();
   render(
     <FreeTrialRadio
+      preview={previewData}
       isUsingFreeTrial={false}
       setIsUsingFreeTrial={setIsUsingFreeTrial}
     />
