+++ static/js/src/PurchaseModal/components/ModalParts/ModalBody.tsx
