+++ static/js/src/PurchaseModal/components/ModalParts/ModalFooter.tsx
