+++ static/js/src/PurchaseModal/components/ModalParts/ModalHeader.tsx
