+++ static/js/src/PurchaseModal/components/PaymentMethodSummary/index.ts
