+++ static/js/src/PurchaseModal/components/TermsCheckbox/TermsCheckbox.test.tsx
