+++ static/js/src/PurchaseModal/components/TermsCheckbox/TermsCheckbox.tsx
