+++ static/js/src/PurchaseModal/components/TermsCheckbox/index.ts
