+++ static/js/src/PurchaseModal/utils/utils.ts
