+++ static/js/src/advantage/subscribe/react/components/BuyButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./BuyButton";
